trigger-source: New source plugin
This is a very simple plugin, that does nothing more than periodically
create a message out of thin air. The purpose of the plugin is to serve
as a very simple source plugin example, and to aid debugging by making
it easy to generate messages on a timer.

Configuration is simple:

 source s_trigger {
  trigger(trigger-freq(10) # the default
          tags("trigger-happy")
          program-override("trigger")
          host-override("trigger-happy")
  );
 };

Signed-off-by: Gergely Nagy <algernon@balabit.hu>