trigger-source: Code cleanup
Instead of having trigger_freq and trigger_message in both TriggerSource
and TriggerSourceDriver (a borrowed reference in the former), have an
inline struct in TriggerSourceDriver, to which a pointer is stored in
TriggerSource.

This way adding new options in the future becomes much easier, as they
will be available to TriggerSource automatically.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>