Makefile.am: Do not ship generated grammar files
The generated grammar files are tied to a particular syslog-ng version,
and are not reusable. The tarball should compile with all supported
versions of syslog-ng, regardless of grammar files.

For this reason, BUILT_SOURCES have been dropped, and nodist_* variables
are used instead. This also means that one needs flex & bison to compile
the Incubator, which is now noted in the README.

This fixes #5.

Reported-by: Peter Czanik <czanik@balabit.hu>
Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>