lua: Added mode parameter to lua destination.
The new mode parameter allows two mode: formatted and raw.
The formatted mode is the "legacy" mode in which we can specify a template
and the queue-func gets a formatted message as a string parameter.
In raw mode, the queue-func gets the message as a Lua table, so the fields
of the message can be accessed directly. For example:

function queue_function(msg)
  print('Message:'..msg['MESSAGE'])
  print('Host:'..msg['HOST'])
end

In raw mode, the template parameter of the destination is omitted. Every
field of the message is accessed as string. (even the numeric ones, eg.: PID).