monitor-source: Added monitoring source.
Monitor source is a new source mixing the capabilities of a Lua-based
driver and the trigger source. It is periodically executing a given
Lua function. If the function returns a string, a new message is created
which contains this string as the MESSAGE part. However, when a (non-nested)
dict is returned, the key/value pairs of the dict is added as name-value pairs
to the newly created message.