lua: Add support for setting global variables
This adds a global() option to the lua destination, where one can list
name-value pairs, which will be injected into the lua state upon
startup, shortly after the script is loaded. Currently, it supports only
strings and 32-bit integers. This fixes #23.

An usage example is included in the elasticsearch example.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>