configure.ac: Support compiling against a syslog-ng with internal ivykis
If syslog-ng was compiled with an internal ivykis, it will (in upcoming
versions) install the ivykis headers and adjust CFLAGS accordingly, so
that modules can be compiled against libsyslog-ng alone. This will be
marked with an "ivykis=internal" in the syslog-ng pkg-config control
file. When that is found, the Incubator need not look for a system-wide
Ivykis installation.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>