date-parser: allow a user to set timezone for timezone-less dates
When the parsed date doesn't come with a timezone, allow a user to set
another timezone using the `time-zone()` parameter. The timezone can be
expressed as `[+-]HH:MM` or as a name (like `Europe/Paris`).

Behavior of `mktime()` is quite undocumented. Tests are mandatory to
detect any problem across supported platforms.

Signed-off-by: Vincent Bernat <vincent@bernat.im>