kafka: when no `--with-librdkafka` flag, check headers
When the user doesn't force Kafka location, check at the regular
place. This was working without that previously because the other branch
was used and because the headers were present at the default location,
the test was working even if additional paths were provided.

We also ensure that `$with_librdkafka` is set to `no` when not provided.

Signed-off-by: Vincent Bernat <vincent@bernat.im>